+++ src/app/company/[id]/manage_camp/[camp_id]/manage_camp_area/_styles/CampAreaForm.module.css
@@ -1,3 +1,40 @@
+/* 캠핑존 페이지 css */
+.backWidth {
+  width: 1200px;
+  margin: 0 auto;
+  padding: 40px 0;
+}
+.campAreaWrap {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-column-gap: 20px;
+  grid-row-gap: 20px;
+}
+.addCampArea {
+  width: 590px;
+  border: 1px solid #333;
+  border-radius: 37px;
+  height: 335px;
+}
+
+/* 캠핑존 모달 css */
+.modalUP {
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  left: 0;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.6);
+  z-index: 1;
+}
+.wrap {
+  width: 300px;
+  height: 450px;
+  background: #fff;
+  margin: 40px auto;
+}
 .formLayout {
   display: grid;
   width: 80%;
